post cont chng

+++ src/main/java/miguelguzman/venusrestblog/controller/PostsController.java
@@ -4,12 +4,16 @@ import lombok.AllArgsConstructor;
 import miguelguzman.venusrestblog.data.Category;
 import miguelguzman.venusrestblog.data.Post;
 import miguelguzman.venusrestblog.data.User;
+import miguelguzman.venusrestblog.misc.FieldHelper;
 import miguelguzman.venusrestblog.repository.CategoriesRepository;
 import miguelguzman.venusrestblog.repository.PostsRepository;
 import miguelguzman.venusrestblog.repository.UsersRepository;
-import miguelguzman.venusrestblog.security.OAuthConfiguration;
 import miguelguzman.venusrestblog.service.EmailService;
+import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 
 import java.util.ArrayList;
@@ -20,6 +24,7 @@ import java.util.Optional;
 @RestController
 @RequestMapping(value = "/api/posts", produces = "application/json")
 public class PostsController {
+    private EmailService emailService;
     private PostsRepository postsRepository;
     private UsersRepository usersRepository;
     private CategoriesRepository categoriesRepository;
@@ -31,65 +36,56 @@ public class PostsController {
 
     @GetMapping("/{id}")
     public Optional<Post> fetchPostById(@PathVariable long id) {
-        return postsRepository.findById(id);
+        Optional<Post> optionalPost = postsRepository.findById(id);
+        if(optionalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post id " + id + " not found");
+        }
+        return optionalPost;
     }
 
     @PostMapping("")
-    public void createPost(@RequestBody Post newPost, OAuthConfiguration auth) {
+    public void createPost(@RequestBody Post newPost, OAuth2Authentication auth) {
         String userName = auth.getName();
         User author = usersRepository.findByUserName(userName);
         newPost.setAuthor(author);
 
-        // use a fake author for the post
-        User author = usersRepository.findById(1L).get();
-        newPost.setAuthor(author);
+        newPost.setCategories(new ArrayList<>());
 
+        // use first 2 categories for the post by default
         Category cat1 = categoriesRepository.findById(1L).get();
         Category cat2 = categoriesRepository.findById(2L).get();
-        newPost.setCategories(new ArrayList<>());
+
         newPost.getCategories().add(cat1);
         newPost.getCategories().add(cat2);
-//
-//        // make some fake categories and throw them in the new post
-//        Category cat1 = new Category(1L, "bunnies", null);
-//        Category cat2 = new Category(2L, "margaritas", null);
-//        newPost.setCategories(new ArrayList<>());
-//        newPost.getCategories().add(cat1);
-//        newPost.getCategories().add(cat2);
 
         postsRepository.save(newPost);
+
+        emailService.prepareAndSend(newPost, "Hey man you made a post", "See subject");
     }
 
     @DeleteMapping("/{id}")
     public void deletePostById(@PathVariable long id) {
+        Optional<Post> optionalPost = postsRepository.findById(id);
+        if(optionalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post id " + id + " not found");
+        }
         postsRepository.deleteById(id);
-        // what to do if we don't find it
-//        throw new RuntimeException("Post not found");
     }
 
     @PutMapping("/{id}")
     public void updatePost(@RequestBody Post updatedPost, @PathVariable long id) {
+        Optional<Post> originalPost = postsRepository.findById(id);
+        if(originalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post " + id + " not found");
+        }
+
         // in case id is not in the request body (i.e., updatedPost), set it
         // with the path variable id
         updatedPost.setId(id);
-        postsRepository.save(updatedPost);
 
-//        // find the post to update in the posts list
-//
-//        Post post = findPostById(id);
-//        if(post == null) {
-//            System.out.println("Post not found");
-//        } else {
-//            if(updatedPost.getTitle() != null) {
-//                post.setTitle(updatedPost.getTitle());
-//            }
-//            if(updatedPost.getContent() != null) {
-//                post.setContent(updatedPost.getContent());
-//            }
-//            return;
-//        }
-//        throw new RuntimeException("Post not found");
+        // copy any new field values FROM updatedPost TO originalPost
+        BeanUtils.copyProperties(updatedPost, originalPost.get(), FieldHelper.getNullPropertyNames(updatedPost));
+
+        postsRepository.save(originalPost.get());
     }
-
-
 }
